Refactor Validator and remove unused Validator classes
