--- src/Core/Validation/PostCreateValidator.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace App\Core\Validation;
-
-use Exception;
-
-final class PostCreateValidator extends Validator implements ValidatorInterface
-{
-    /**
-     * @throws Exception
-     *
-     * @return $this
-     */
-    public function getValidator(): self
-    {
-        foreach ($this->data as $key => $value) {
-            $this->required($key)
-                ->notBlank($key)
-            ;
-        }
-
-        return $this;
-    }
-}
