--- src/Core/Validation/ValidatorInterface.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Core\Validation;
-
-interface ValidatorInterface
-{
-    /**
-     * @return $this
-     */
-    public function getValidator(): self;
-}
